# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "theclub-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/roulette/rounds": {
      "post": {
        "tags": [
          "roulette"
        ],
        "summary": "Create Round",
        "operationId": "create_round_api_v1_roulette_rounds_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceRoundRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoundResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "roulette"
        ],
        "summary": "List Rounds",
        "operationId": "list_rounds_api_v1_roulette_rounds_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoundHistoryPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BetInput": {
        "properties": {
          "bet_type": {
            "$ref": "#/components/schemas/BetType"
          },
          "selection": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                }
              ]
            },
            "type": "object",
            "title": "Selection"
          },
          "stake_minor": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Stake Minor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bet_type",
          "selection",
          "stake_minor"
        ],
        "title": "BetInput"
      },
      "BetType": {
        "type": "string",
        "enum": [
          "straight",
          "split",
          "street",
          "corner",
          "line",
          "dozen",
          "column",
          "red",
          "black",
          "odd",
          "even",
          "high",
          "low"
        ],
        "title": "BetType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "PlaceRoundRequest": {
        "properties": {
          "bets": {
            "items": {
              "$ref": "#/components/schemas/BetInput"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Bets"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bets"
        ],
        "title": "PlaceRoundRequest"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest"
      },
      "RoundHistoryItem": {
        "properties": {
          "round_id": {
            "type": "string",
            "format": "uuid",
            "title": "Round Id"
          },
          "outcome": {
            "type": "integer",
            "title": "Outcome"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "bets": {
            "items": {
              "$ref": "#/components/schemas/SettledBetResponse"
            },
            "type": "array",
            "title": "Bets"
          }
        },
        "type": "object",
        "required": [
          "round_id",
          "outcome",
          "created_at",
          "bets"
        ],
        "title": "RoundHistoryItem"
      },
      "RoundHistoryPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RoundHistoryItem"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "RoundHistoryPage"
      },
      "RoundResponse": {
        "properties": {
          "round_id": {
            "type": "string",
            "format": "uuid",
            "title": "Round Id"
          },
          "outcome": {
            "type": "integer",
            "title": "Outcome"
          },
          "bets": {
            "items": {
              "$ref": "#/components/schemas/SettledBetResponse"
            },
            "type": "array",
            "title": "Bets"
          },
          "total_stake_minor": {
            "type": "integer",
            "title": "Total Stake Minor"
          },
          "total_payout_minor": {
            "type": "integer",
            "title": "Total Payout Minor"
          },
          "net_minor": {
            "type": "integer",
            "title": "Net Minor"
          },
          "balance_minor": {
            "type": "integer",
            "title": "Balance Minor"
          }
        },
        "type": "object",
        "required": [
          "round_id",
          "outcome",
          "bets",
          "total_stake_minor",
          "total_payout_minor",
          "net_minor",
          "balance_minor"
        ],
        "title": "RoundResponse"
      },
      "SettledBetResponse": {
        "properties": {
          "bet_type": {
            "$ref": "#/components/schemas/BetType"
          },
          "selection": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                }
              ]
            },
            "type": "object",
            "title": "Selection"
          },
          "stake_minor": {
            "type": "integer",
            "title": "Stake Minor"
          },
          "payout_minor": {
            "type": "integer",
            "title": "Payout Minor"
          },
          "won": {
            "type": "boolean",
            "title": "Won"
          }
        },
        "type": "object",
        "required": [
          "bet_type",
          "selection",
          "stake_minor",
          "payout_minor",
          "won"
        ],
        "title": "SettledBetResponse"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "const": "bearer",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}